requirements.txt
langchain
langchain-community
langchain_core
langchain-openai
langchain-qdrant
langchain_huggingface
langchain-chroma
langchain-postgres
langchain-pinecone
langgraph
langgraph-checkpoint-sqlite
openpyxl
fastapi
fastembed
uvicorn
pandas
pydantic
python-dotenv
qdrant-client
sentence-transformers
tiktoken
rapidfuzz
aiosqlite
jq
deepeval
openpyxl
ragas
weaviate
weaviate-client
chroma
psycopg2-binary
sqlalchemy
psycopg[binary]
pinecone-client
